sna: Extract correct sna_event struct in vblank handler
When extracting the underlying struct sna object from the drm_vblank_event,
we need to look at the event type. If the event is actually from a DRI3/present
call 2nd bit of the event->user_data will be set high. In that case
we need to first extract the sna_present_event, and then we can get the
struct sna pointer from the event.

This fixes a crash introduced by the previous commit.
